Pre-Alpha v0.005b
Made sure the grass communicates better, the player can now also slide
and wall slide. And I connected the ends of the world, so if you walk
out of the left side of the level, you end up at the right side. And
some changes to the level were made.